Rename to async.ts
